# server/openapi.yaml
openapi: 3.0.3
info:
  title: BaudAgain BBS API
  description: |
    REST API for BaudAgain BBS - A modern AI-enhanced bulletin board system.
    
    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your_jwt_token>
    ```
    
    ## Rate Limiting
    - Authentication endpoints: 10 requests per minute
    - Read operations: 100 requests per 15 minutes
    - Write operations: 30 requests per minute
    - Message posting: 30 messages per hour
    
  version: 1.0.0
  contact:
    name: BaudAgain BBS
  license:
    name: MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              enum: [INVALID_INPUT, UNAUTHORIZED, FORBIDDEN, NOT_FOUND, CONFLICT, RATE_LIMIT_EXCEEDED, INTERNAL_ERROR]
            message:
              type: string
            details:
              type: object
      required:
        - error
...
    Door:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        category:
          type: string
        accessLevel:
          type: integer
        canAccess:
          type: boolean
...
paths:
...
  /doors:
    get:
      tags:
        - Doors
      summary: List doors
      description: Get available door games
      operationId: listDoors
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of doors
          content:
            application/json:
              schema:
                type: object
                properties:
                  doors:
                    type: array
                    items:
                      $ref: '#/components/schemas/Door'
  /doors/{id}/enter:
    post:
      tags:
        - Doors
      summary: Enter door
      description: |
        Enter a door game and create a session
        
        **curl Example:**
        ```bash
        curl -X POST "http://localhost:8080/api/v1/doors/the_oracle/enter" \
          -H "Authorization: Bearer $TOKEN"
        ```
      operationId: enterDoor
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Door game ID
      responses:
        '200':
          description: Door entered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionId:
                    type: string
                    format: uuid
                    description: Session ID for subsequent requests
                  doorId:
                    type: string
                    description: Door game ID
                  doorName:
                    type: string
                    description: Door game name
                  output:
                    type: string
                    description: Initial door output (ANSI formatted)
                  resumed:
                    type: boolean
                    description: Whether this is a resumed session from the database
        '404':
          description: Door not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Failed to enter door
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /doors/{id}/input:
    post:
      tags:
        - Doors
      summary: Send door input
      description: |
        Send user input to an active door game session
        
        **curl Example:**
        ```bash
        curl -X POST "http://localhost:8080/api/v1/doors/the_oracle/input" \
          -H "Authorization: Bearer $TOKEN" \
          -H "Content-Type: application/json" \
          -d '{
            "input": "What is the meaning of life?"
          }'
        ```
      operationId: sendDoorInput
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Door game ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - input
              properties:
                sessionId:
                  type: string
                  format: uuid
                  description: Session ID (optional if using same user session)
                input:
                  type: string
                  description: User input to send to the door
      responses:
        '200':
          description: Input processed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionId:
                    type: string
                    format: uuid
                    description: Session ID
                  output:
                    type: string
                    description: Door response output (ANSI formatted)
                  exited:
                    type: boolean
                    description: Whether the user has exited the door
        '400':
          description: Invalid input or state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Session does not belong to current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Door or session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Failed to process input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /doors/{id}/exit:
    post:
      tags:
        - Doors
      summary: Exit door
      description: Exit an active door game session
      operationId: exitDoor
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Door game ID
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                sessionId:
                  type: string
                  format: uuid
                  description: Session ID (optional if using same user session)
      responses:
        '200':
          description: Door exited successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  output:
                    type: string
                    description: Exit message (ANSI formatted)
                  exited:
                    type: boolean
                    description: Always true
        '403':
          description: Session does not belong to current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Door or session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Failed to exit door
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /doors/{id}/session:
    get:
      tags:
        - Doors
      summary: Get door session info
      description: Get information about the current user's session in a specific door
      operationId: getDoorSession
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Door game ID
      responses:
        '200':
          description: Door session information
          content:
            application/json:
              schema:
                type: object
                properties:
                  inDoor:
                    type: boolean
                    description: Whether user is currently in this door
                  hasSavedSession:
                    type: boolean
                    description: Whether there is a saved session in the database
                  sessionId:
                    type: string
                    format: uuid
                    nullable: true
                    description: Session ID if in door
                  doorId:
                    type: string
                    description: Door game ID
                  doorName:
                    type: string
                    description: Door game name
                  lastActivity:
                    type: string
                    format: date-time
                    description: Last activity timestamp
                  gameState:
                    type: object
                    description: Current game state (only if inDoor is true)
                  history:
                    type: array
                    description: Conversation history (only if inDoor is true)
                    items:
                      type: object
                  savedState:
                    type: object
                    description: Saved game state from database (only if hasSavedSession is true)
                  savedHistory:
                    type: array
                    description: Saved conversation history from database (only if hasSavedSession is true)
                    items:
                      type: object
        '404':
          description: Door not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /doors/{id}/resume:
    post:
      tags:
        - Doors
      summary: Resume saved door session
      description: Resume a previously saved door game session from the database
      operationId: resumeDoorSession
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Door game ID
      responses:
        '200':
          description: Door session resumed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionId:
                    type: string
                    format: uuid
                    description: Session ID for subsequent requests
                  doorId:
                    type: string
                    description: Door game ID
                  doorName:
                    type: string
                    description: Door game name
                  output:
                    type: string
                    description: Door output with resumed state (ANSI formatted)
                  resumed:
                    type: boolean
                    description: Always true for this endpoint
        '404':
          description: Door not found or no saved session exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Failed to resume door session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /doors/my-sessions:
    get:
      tags:
        - Doors
      summary: Get user's saved door sessions
      description: Get all saved door sessions for the current user
      operationId: getMySavedDoorSessions
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of saved door sessions
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessions:
                    type: array
                    items:
                      type: object
                      properties:
                        doorId:
                          type: string
                          description: Door game ID
                        doorName:
                          type: string
                          description: Door game name
                        lastActivity:
                          type: string
                          format: date-time
                          description: Last activity timestamp
                        createdAt:
                          type: string
                          format: date-time
                          description: Session creation timestamp
                        canResume:
                          type: boolean
                          description: Whether this session can be resumed
                  totalCount:
                    type: integer
                    description: Total number of saved sessions
  /doors/sessions:
    get:
      tags:
        - Doors
      summary: Get all active door sessions
      description: Get information about all active door sessions (admin only)
      operationId: getAllDoorSessions
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of active door sessions
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessions:
                    type: array
                    items:
                      type: object
                      properties:
                        sessionId:
                          type: string
                          format: uuid
                        userId:
                          type: string
                          format: uuid
                        handle:
                          type: string
                        doorId:
                          type: string
                        doorName:
                          type: string
                        lastActivity:
                          type: string
                          format: date-time
                        inactiveTime:
                          type: integer
                          description: Milliseconds since last activity
                  totalCount:
                    type: integer
                    description: Total number of active door sessions
        '403':
          description: Insufficient permissions (admin only)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /doors/{id}/stats:
    get:
      tags:
        - Doors
      summary: Get door statistics
      description: Get statistics about a specific door game
      operationId: getDoorStats
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Door game ID
      responses:
        '200':
          description: Door statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  doorId:
                    type: string
                  doorName:
                    type: string
                  activeSessions:
                    type: integer
                    description: Number of users currently in this door
                  timeout:
                    type: integer
                    description: Door timeout in milliseconds
        '404':
          description: Door not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...